# openapi.yaml
openapi: "3.1.0"

info:
  title: UIDE Inscripciones API
  version: "2.0.0"
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://localhost:3000
    description: Desarrollo local

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Estado del servidor
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Servidor activo.
        '401':
          description: API key ausente o inválida.

  /v1/inscripciones:
    post:
      operationId: crearInscripcionV1
      summary: Crear inscripción contrato v1
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscripcionV1'
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440000"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"
      responses:
        '201':
          description: Inscripción creada.
        '400':
          description: Payload inválido.
        '401':
          description: API key ausente o inválida.

  /v2/inscripciones:
    post:
      operationId: crearInscripcionV2
      summary: Crear inscripción con método de pago contrato v2
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscripcionV2'
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440000"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"
              payment_method: "scholarship"
      responses:
        '201':
          description: Inscripción creada.
        '400':
          description: Payload inválido o payment_method fuera de enum.
        '401':
          description: API key ausente o inválida.

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    InscripcionV1:
      type: object
      required:
        - estudianteId
        - materias
        - periodoId
      properties:
        estudianteId:
          type: string
          format: uuid
        materias:
          type: array
          minItems: 1
          items:
            type: string
        periodoId:
          type: string

    InscripcionV2:
      allOf:
        - $ref: '#/components/schemas/InscripcionV1'
        - type: object
          required:
            - payment_method
          properties:
            payment_method:
              type: string
              enum:
                - debit
                - credit
                - scholarship
      unevaluatedProperties: false